+++ StopSignProject/app/src/main/java/cn/truthvision/stopsignproject/Video.java
@@ -19,9 +19,6 @@ import com.google.android.gms.common.api.GoogleApiClient;
 import cn.truthvision.stopsignlib.VideoManager;
 
 public class Video extends AppCompatActivity {
-    private static final int CAPTURE_VIDEO_ACTIVITY_REQUEST_CODE = 200;
-    private static final int CAPTURE_IMAGE_ACTIVITY_REQUEST_CODE = 100;
-    private Uri fileUri;
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
@@ -30,55 +27,6 @@ public class Video extends AppCompatActivity {
         //setSupportActionBar(toolbar);
 
         Intent i = getIntent();
-
-
-
-
-        /*Note: I will be editing the video portion so please do not touch the code between these comments
-        **************************************************************************************************
-        **************************************************************************************************
-        **************************************************************************************************
-        **************************************************************************************************
-        **************************************************************************************************
-        **************************************************************************************************
-        **************************************************************************************************
-        **************************************************************************************************
-        **************************************************************************************************
-        **************************************************************************************************
-         */
-
-
-
-
-        //create new Intent
-        Intent intent = new Intent(MediaStore.ACTION_VIDEO_CAPTURE);
-
-        int MEDIA_TYPE_VIDEO = 2;
-
-        fileUri = VideoManager.getOutputMediaFileUri(MEDIA_TYPE_VIDEO);  // create a file to save the video
-        intent.putExtra(MediaStore.EXTRA_OUTPUT, fileUri);  // set the image file name
-
-        intent.putExtra(MediaStore.EXTRA_VIDEO_QUALITY, 1); // set the video image quality to high
-
-        // start the Video Capture Intent
-        startActivityForResult(intent, CAPTURE_VIDEO_ACTIVITY_REQUEST_CODE);
-
-
-
-        /*
-        **************************************************************************************************
-        **************************************************************************************************
-        **************************************************************************************************
-        **************************************************************************************************
-        **************************************************************************************************
-        **************************************************************************************************
-        **************************************************************************************************
-        **************************************************************************************************
-        **************************************************************************************************
-        **************************************************************************************************
-         */
-
-
 
         Button settings = (Button) findViewById(R.id.settingsbutton2);
         settings.setOnClickListener(new View.OnClickListener() {
@@ -121,9 +69,10 @@ public class Video extends AppCompatActivity {
     }
 
     public void onStartRecording(View view) {
+        Intent i = new Intent(this, Recording.class);
+        startActivity(i);
 
     }
-
 
 
     public void onData(View view) {
@@ -132,31 +81,4 @@ public class Video extends AppCompatActivity {
         finish();
     }
 
-
-    //Intents for Video
-    protected void onActivityResult(int requestCode, int resultCode, Intent data) {
-        if (requestCode == CAPTURE_IMAGE_ACTIVITY_REQUEST_CODE) {
-            if (resultCode == RESULT_OK) {
-                // Image captured and saved to fileUri specified in the Intent
-                Toast.makeText(this, "Image saved to:\n" +
-                        data.getData(), Toast.LENGTH_LONG).show();
-            } else if (resultCode == RESULT_CANCELED) {
-                // User cancelled the image capture
-            } else {
-                // Image capture failed, advise user
-            }
-        }
-
-        if (requestCode == CAPTURE_VIDEO_ACTIVITY_REQUEST_CODE) {
-            if (resultCode == RESULT_OK) {
-                // Video captured and saved to fileUri specified in the Intent
-                Toast.makeText(this, "Video saved to:\n" +
-                        data.getData(), Toast.LENGTH_LONG).show();
-            } else if (resultCode == RESULT_CANCELED) {
-                // User cancelled the video capture
-            } else {
-                // Video capture failed, advise user
-            }
-        }
-    }
 }
