+++ StopSignProject/stopsignlib/src/main/java/cn/truthvision/stopsignlib/VideoManager.java
@@ -26,14 +26,14 @@ public class VideoManager {
         // using Environment.getExternalStorageState() before doing this.
 
         File mediaStorageDir = new File(Environment.getExternalStoragePublicDirectory(
-                Environment.DIRECTORY_PICTURES), "MyCameraApp");
+                Environment.DIRECTORY_PICTURES), "StopSignVidStore");
         // This location works best if you want the created images to be shared
         // between applications and persist after your app has been uninstalled.
 
         // Create the storage directory if it does not exist
         if (!mediaStorageDir.exists()) {
             if (!mediaStorageDir.mkdirs()) {
-                Log.d("MyCameraApp", "failed to create directory");
+                Log.d("StopSignVidStore", "failed to create directory");
                 return null;
             }
         }
